Refactor: Move RegexValidator from bisq.desktop.util.validation to bisq.core.util.validation
Add RegexValidatorFactory
Move regex tests from GUIUtilTest to new RegexValidatorTest class
