+++ core/src/main/java/bisq/core/util/validation/RegexValidator.java
@@ -1,9 +1,9 @@
-package bisq.desktop.util.validation;
+package bisq.core.util.validation;
 
 import bisq.core.locale.Res;
-import bisq.core.util.validation.InputValidator;
 
 public class RegexValidator extends InputValidator {
+
     private String pattern;
     private String errorMessage;
